endpoints/file-endpoint/requirements.txt
Pillow==10.1.0
lxml==5.1.0
